Worked around a Win10 RS2 bug that impacts hit testing on transparent windows
Windows 10 RS2 introduced a bug where hit testing breaks on this and
other similar apps where there is a transparent window. We spawn a
transparent window so that you can do MouseDown/MouseMove events to get
a snip ready (and we draw the red rectangle so you know the area you are
about to capture). And then when you do MouseUp we take screenshot at
that moment..

On RS2 this does not work, so I added a workaround where we screenshot
the desktop as soon as you go into snipping mode. So you end up snipping
a still image.

The workaround is really only a worse experience when snipping video
content as the end user will not be able to see the video play while in
snipping mode.

Hopefully this bug is fixed in Windows 10 RS3 and we can go back to
normal.